+++ App/Sources/Enum/MessageType.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+enum MessageType: String, CaseIterable {
+    case normal = "BASIC"
+    case password = "CODE"
+}
