+++ App/Sources/Enum/ProgressType.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+enum ProgressType: Int, CaseIterable {
+    case type1 = 1
+    case type2 = 2
+    case type3 = 3
+    case type4 = 4
+    case type5 = 5
+    case type6 = 6
+    case type7 = 7
+}
