+++ App/Sources/Enum/TransportationType.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+enum TransportationType: String, CaseIterable {
+    case airplane = "PLAIN"
+    case car = "CAR"
+    case bicycle = "BICYCLE"
+    case horse = "HORSE"
+    case running = "RUN"
+    case walk = "WALK"
+}
